fix: compress move under ziphelper
